pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "ai-book-converter"
version = "0.1.0"
description = "Project for OCR - books to Epub conversion"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markdown>=3.10",
    "mistralai>=1.10.0",
    "pytest>=9.0.2",
    "pytest-dotenv>=0.5.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.7",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_book_converter"]

[project.scripts]
ai-book-converter = "ai_book_converter.cli:main"

[tool.pytest.ini_options]
env_files = [
    ".env",
]
